Moved theme-settings
