+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/theme-settings/index.tsx
